move app db health check to DataAccess project, fixes executor proxy native AOT build
