+++ Biflow.Aspire.ServiceDefaults/Biflow.Aspire.ServiceDefaults.csproj
@@ -10,7 +10,6 @@
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="JetBrains.Annotations" />
-    <PackageReference Include="Microsoft.Data.SqlClient" />
     <PackageReference Include="Microsoft.Extensions.Http.Resilience" />
     <PackageReference Include="Microsoft.Extensions.ServiceDiscovery"/>
     <PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" />
