+++ Biflow.Ui/Program.cs
@@ -10,7 +10,8 @@ using Biflow.Ui.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.AddServiceDefaults(addDbHealthCheck: true);
+builder.AddServiceDefaults()
+    .AddDatabaseHealthCheck();
 
 builder.Services.AddWindowsService();
 builder.Services.AddSystemd();
